// types/Collections.ts
export enum Collection {
  STEAM_PUNKS = 'Steam Punks',
  ASGARDIANS = 'The Asgardians',
  REAVERS = 'Reavers',
  LAST_HAVEN = 'Last Haven',
  BROHALLA = 'BroHalla',
  ELEMENTERRA_RABBITS = 'Elementerra Rabbits',
  ATOMS = 'Dragon Sols',
  DRAGON_SOLS = 'Dragon Sols',
  SIRENS = 'The Seven Seas: Sirens',
  SEVEN_SEAS = 'The Seven Seas',
  GENESIS_SHIPS = 'Genesis Ships',
}